# specs/003-003-auth-security/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App Authentication & Security API
  description: |
    JWT-based authentication and authorization API for the Todo application.

    This API provides secure user registration, login, and JWT token management
    for multi-user task management with database-level user isolation.

    **Authentication Flow**:
    1. User registers with email and password (POST /api/auth/signup)
    2. User logs in with credentials (POST /api/auth/signin)
    3. Backend returns JWT token with 7-day expiration
    4. Frontend attaches token to all API requests (Authorization: Bearer <token>)
    5. Backend verifies token and extracts user_id for database queries

    **Security Features**:
    - Password hashing with bcrypt (cost factor 10)
    - JWT tokens signed with HS256 algorithm
    - Token expiration after 7 days
    - User isolation at database query level
    - 404 (not 403) for ownership violations
  version: 1.0.0
  contact:
    name: Todo App Team
    email: support@todoapp.example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todoapp.example.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: |
        Create a new user account with email and password.

        **Validation Rules**:
        - Email must be valid format (RFC 5322)
        - Email must be unique (not already registered)
        - Password must be at least 8 characters

        **Security**:
        - Password is hashed with bcrypt (cost factor 10) before storage
        - Plain text password is never stored in database
      operationId: signup
      security: []  # No authentication required for signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            examples:
              valid:
                summary: Valid signup request
                value:
                  email: user@example.com
                  password: securepassword123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                success:
                  summary: Successful signup
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    email: user@example.com
                    created_at: '2026-01-15T10:30:00Z'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_email:
                  summary: Invalid email format
                  value:
                    code: INVALID_EMAIL
                    message: Invalid email format
                    details: {}
                weak_password:
                  summary: Password too short
                  value:
                    code: WEAK_PASSWORD
                    message: Password must be at least 8 characters
                    details: {}
                duplicate_email:
                  summary: Email already registered
                  value:
                    code: DUPLICATE_EMAIL
                    message: Email already registered
                    details: {}
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticate user with email and password, return JWT token.

        **Authentication Process**:
        1. Verify email exists in database
        2. Verify password matches hashed password (bcrypt)
        3. Generate JWT token with user_id and email claims
        4. Sign token with HS256 algorithm using JWT_SECRET
        5. Set expiration to 7 days from now
        6. Return token to frontend

        **Token Usage**:
        - Frontend stores token in localStorage or httpOnly cookie
        - Frontend attaches token to all API requests in Authorization header
        - Token format: `Authorization: Bearer <token>`
      operationId: signin
      security: []  # No authentication required for signin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
            examples:
              valid:
                summary: Valid signin request
                value:
                  email: user@example.com
                  password: securepassword123
      responses:
        '200':
          description: Login successful, JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              examples:
                success:
                  summary: Successful login
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiZXhwIjoxNzM3ODQ5NjAwLCJpYXQiOjE3MzcyNDQ4MDB9.signature_here
                    token_type: bearer
                    expires_in: 604800
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_credentials:
                  summary: Invalid email or password
                  value:
                    code: INVALID_CREDENTIALS
                    message: Invalid email or password
                    details: {}
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rate_limit:
                  summary: Rate limit exceeded
                  value:
                    code: RATE_LIMIT_EXCEEDED
                    message: Too many login attempts. Please try again later.
                    details:
                      retry_after: 60
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /api/auth/signin endpoint.

        **Format**: `Authorization: Bearer <token>`

        **Token Structure**:
        - Header: `{ "alg": "HS256", "typ": "JWT" }`
        - Payload: `{ "user_id": "UUID", "email": "string", "exp": timestamp, "iat": timestamp }`
        - Signature: HMACSHA256(header + payload, JWT_SECRET)

        **Expiration**: 7 days (604800 seconds)
  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address (must be unique)
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: User's password (min 8 characters)
          example: securepassword123
    SigninRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User's password
          example: securepassword123
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiZXhwIjoxNzM3ODQ5NjAwLCJpYXQiOjE3MzcyNDQ4MDB9.signature_here
        token_type:
          type: string
          enum: [bearer]
          description: Token type (always 'bearer')
          example: bearer
        expires_in:
          type: integer
          description: Token expiration in seconds (7 days)
          example: 604800
    UserResponse:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: User's unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2026-01-15T10:30:00Z'
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code (e.g., INVALID_TOKEN, TOKEN_EXPIRED)
          example: INVALID_TOKEN
        message:
          type: string
          description: Human-readable error message
          example: Invalid token
        details:
          type: object
          additionalProperties: true
          description: Additional error details
          example: {}
...